# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Auth",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
...
        "/api/espacios": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Obtener una lista de espacios",
                "description": "Retorna una lista de espacios filtrados por tipo, capacidad y disponibilidad en un rango de fechas.",
                "operationId": "5b30a51c97d21af853cacca0c49ab978",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Filtra los espacios por tipo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "capacidad",
                        "in": "query",
                        "description": "Filtra los espacios con capacidad mínima especificada",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fecha_inicio",
                        "in": "query",
                        "description": "Fecha de inicio para filtrar la disponibilidad",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "query",
                        "description": "Fecha de fin para filtrar la disponibilidad",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de espacios filtrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Espacio"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token not provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Crear un nuevo espacio",
                "description": "Crea un nuevo espacio con los datos proporcionados.",
                "operationId": "d03ef5ee956c549c1a8371ae4d24fc39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Espacio"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Espacio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Espacio"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token not provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/espacios/{id}": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Obtener detalles de un espacio",
                "description": "Retorna los detalles de un espacio específico basado en su ID.",
                "operationId": "e426367fc6788a5cc54af7230faa86ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Espacio"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Espacio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Actualizar un espacio",
                "description": "Actualiza los datos de un espacio existente.",
                "operationId": "3ada5c4d5167b627fced5871ab47aa45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Espacio"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Espacio actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Espacio"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token not provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Espacio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Eliminar un espacio",
                "description": "Elimina un espacio basado en su ID.",
                "operationId": "4cfb60fb59c3e4c3758125082782c32d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Espacio eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token not provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Espacio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Espacio": {
                "title": "Espacio",
                "required": [
                    "tipo",
                    "capacidad"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Sala de Conferencias"
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 50
                    },
                    "ubicacion": {
                        "type": "string",
                        "example": "Edificio A, Piso 3"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-10T10:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}